GOB.SH the IT BOT

+++ snake/snake.h
@@ -1,10 +1,10 @@
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
 
-extern int go_api_stdout(void*, int);
-extern int go_api_stderr(void*, int);
-extern void* go_api_stdinBuffer(int);
-extern int go_api_stdinLen();
+extern Py_ssize_t go_api_stdout(void*, Py_ssize_t);
+extern Py_ssize_t go_api_stderr(void*, Py_ssize_t);
+extern void* go_api_stdinBuffer(Py_ssize_t);
+extern Py_ssize_t go_api_stdinLen();
 extern void go_api_free(void*);
 
 // PyCFunction signature (both the tuple and the dictionary)
